Allow specifying a custom delimiter for sequence fields
This allows adding a `#[recap(delimiter_regex = "...")]` attribute to a field with a sequence-type value which is used when splitting the string. Currently the string is always split on `","`, but sometimes lists aren't comma joined. This adds the ability to split strings with an arbitrary regex to handle more situations. This was inspired by https://adventofcode.com/2023/day/4, where splitting on spaces is useful. Note that this was also exposed via the functional bring-your-own-captures API in the form of a new `recap::from_captures_with_options` method.

Note that this also refactors the derive code to share a common `RE` instance rather than duplicating it in each method that needs access to the regex. This common location is also where the new field options are stored.

This also adds documentation for the added attribute and a doctest to both demonstrate and verify the behavior.

Since this is a new attribute and behavior, it should probably be added to the changelog.